Add files via upload
Very basic script to use find command to generate a text file that holds the paths of all the flac files in the current directory and descendants, this is fed to the flac-pipe.sh script than then reads the input into the flac.sh script that calls the flac encoder on each flac file and keeps the smaller of the two output files... apparently this nonsense was faster on all my systems than running find and having it directly call the flac.sh script... by about an order of magnitude. The reason they are three scripts is so I can use the pipe script to handle output from programs like find directly if I want or I can just point flac.sh at the flac files in the current directory (or any other files I give it), or I can have the finder script (this one) find the files, make a text file for me to see if it go them all even it is running in the background via nohup, and then re-encode them without my presence being needed... or even for a GUI program to be running which is great for headless machines.